requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
openai>=1.10.0
python-multipart==0.0.6
pydub==0.25.1
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.8.0

# IndexTTS2 依赖
torch>=2.0.0
torchaudio
numpy
librosa
scipy
huggingface_hub
transformers
accelerate
gradio_client>=0.8.0

# Chatterbox Multilingual TTS (可选，支持23种语言)
# 如需使用，请取消注释以下行：
# chatterbox-tts>=1.0.0

# RAG 知识库依赖
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.5
chromadb==0.4.22
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
pypdf==3.17.4
python-docx==1.1.0
markdown==3.5.2
tiktoken==0.5.2

# 质量评估依赖
textstat==0.7.3
aiofiles==23.2.1

# Vision 图片分析依赖
# torch和transformers已在上面
aiohttp==3.9.1
Pillow>=10.1.0
